refactor: conform 공통 유틸함수로 분리
