+++ src/app/events/[eventId]/edit/form.type.ts
@@ -4,7 +4,7 @@ import {
   EventTypeEnum,
   UpdateEventRequest,
 } from '@/types/event.type';
-import { convertToUTC } from '@/utils/date.util';
+import { conformEventData } from '../../conform.util';
 
 export const EditEventFormSchema = z.object({
   status: EventStatusEnum,
@@ -26,17 +26,5 @@ export const EditEventFormSchema = z.object({
 export type EditEventFormData = z.infer<typeof EditEventFormSchema>;
 
 export const conform = (data: EditEventFormData): UpdateEventRequest => {
-  const artistIds = data.artistIds.filter(
-    (artist): artist is { artistId: string } => artist.artistId !== null,
-  );
-  const dailyEvents = data.dailyEvents.map((daily) => ({
-    ...daily,
-    date: convertToUTC(daily.date),
-  }));
-  console.log('in conform dailyEvents', dailyEvents);
-  return {
-    ...data,
-    artistIds: artistIds.map((artist) => artist.artistId),
-    dailyEvents,
-  } satisfies UpdateEventRequest;
+  return conformEventData<EditEventFormData, UpdateEventRequest>(data);
 };
